refactor: service Layer 2
