+++ src/main/java/com/example/book_your_seat/payment/service/command/PaymentCommandService.java
@@ -1,4 +1,4 @@
-package com.example.book_your_seat.payment.service;
+package com.example.book_your_seat.payment.service.command;
 
 import com.example.book_your_seat.payment.domain.Payment;
 import com.example.book_your_seat.payment.repository.PaymentRepository;
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
-public class PaymentCommandServiceImpl {
+public class PaymentCommandService {
 
     private final PaymentRepository paymentRepository;
