+++ src/main/java/com/example/book_your_seat/payment/service/facade/PaymentFacade.java
@@ -1,4 +1,4 @@
-package com.example.book_your_seat.payment.service;
+package com.example.book_your_seat.payment.service.facade;
 
 import com.example.book_your_seat.concert.controller.dto.ConcertResponse;
 import com.example.book_your_seat.concert.service.ConcertQueryService;
@@ -11,7 +11,7 @@ import com.example.book_your_seat.payment.controller.dto.response.ConfirmRespons
 import com.example.book_your_seat.payment.controller.dto.response.FinalPriceResponse;
 import com.example.book_your_seat.payment.domain.Payment;
 import com.example.book_your_seat.payment.domain.PaymentStatus;
-import com.example.book_your_seat.payment.service.PaymentCommandService;
+import com.example.book_your_seat.payment.service.command.PaymentCommandService;
 import com.example.book_your_seat.payment.service.dto.PaymentCommand;
 import com.example.book_your_seat.queue.service.facade.QueueService;
 import com.example.book_your_seat.reservation.domain.Reservation;
@@ -32,7 +32,7 @@ import java.util.List;
 @Transactional
 @RequiredArgsConstructor
 @Service
-public class PaymentFacadeImpl {
+public class PaymentFacade {
 
     private final AddressQueryService addressQueryService;
